refactor: jit actions
